Re #7471 edited design of APIs based on consultation
