+++ server/src/main/java/cz/zcu/yamanager/dto/RequestType.java
@@ -1,5 +1,5 @@
 package cz.zcu.yamanager.dto;
 
 public enum  RequestType {
-    VACATION, AUTHORIZATION
+    VACATION, SICK_DAY, AUTHORIZATION
 }
